# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ZBX Chain Explorer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /blocks:
    get:
      operationId: getBlocks
      tags: [blocks]
      summary: List recent blocks
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of blocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Block"
  /blocks/{blockNumber}:
    get:
      operationId: getBlock
      tags: [blocks]
      summary: Get block by number
      parameters:
        - name: blockNumber
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Block detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockDetail"
        "404":
          description: Block not found
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: List recent transactions
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: type
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /transactions/{hash}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction by hash
      parameters:
        - name: hash
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Transaction detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionDetail"
        "404":
          description: Transaction not found
  /validators:
    get:
      operationId: getValidators
      tags: [validators]
      summary: List all validators
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [active, inactive, jailed] }
      responses:
        "200":
          description: Validator list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Validator"
...
  /validators/{address}:
    get:
      operationId: getValidator
      tags: [validators]
      summary: Get validator details
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Validator detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorDetail"
        "404":
          description: Validator not found
...
  /governance/proposals:
    get:
      operationId: getProposals
      tags: [governance]
      summary: List governance proposals (ZEPs)
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [active, passed, rejected, pending] }
      responses:
        "200":
          description: Proposals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
  /governance/proposals/{id}:
    get:
      operationId: getProposal
      tags: [governance]
      summary: Get ZEP proposal detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Proposal detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProposalDetail"
...
  /payid/resolve/{payid}:
    get:
      operationId: resolvePayId
      tags: [payid]
      summary: Resolve a PayID to an address
      parameters:
        - name: payid
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Resolved address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayIdResolution"
        "404":
          description: PayID not found
...
components:
  schemas:
...
    Block:
      type: object
      required: [number, hash, timestamp, txCount, proposer, gasUsed, gasLimit, size]
      properties:
        number: { type: integer }
        hash: { type: string }
        timestamp: { type: string }
        txCount: { type: integer }
        proposer: { type: string }
        gasUsed: { type: string }
        gasLimit: { type: string }
        size: { type: integer }
        parentHash: { type: string }
        stateRoot: { type: string }
    BlockDetail:
      type: object
      required: [number, hash, timestamp, txCount, proposer, gasUsed, gasLimit, size, parentHash, stateRoot, transactionsRoot, receiptsRoot, difficulty, totalDifficulty, nonce]
      properties:
        number: { type: integer }
        hash: { type: string }
        timestamp: { type: string }
        txCount: { type: integer }
        proposer: { type: string }
        gasUsed: { type: string }
        gasLimit: { type: string }
        size: { type: integer }
        parentHash: { type: string }
        stateRoot: { type: string }
        transactionsRoot: { type: string }
        receiptsRoot: { type: string }
        difficulty: { type: string }
        totalDifficulty: { type: string }
        nonce: { type: string }
        extraData: { type: string }
        baseFeePerGas: { type: string }
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
    Transaction:
      type: object
      required: [hash, blockNumber, timestamp, from, status, value, gasPrice, gasUsed, type]
      properties:
        hash: { type: string }
        blockNumber: { type: integer }
        timestamp: { type: string }
        from: { type: string }
        to: { type: ["string", "null"] }
        status: { type: string, enum: [success, failed, pending] }
        value: { type: string }
        gasPrice: { type: string }
        gasUsed: { type: string }
        type: { type: string }
        nonce: { type: integer }
    TransactionDetail:
      type: object
      required: [hash, blockNumber, timestamp, from, status, value, gasPrice, gasUsed, type, nonce, input, logs]
      properties:
        hash: { type: string }
        blockNumber: { type: integer }
        timestamp: { type: string }
        from: { type: string }
        to: { type: ["string", "null"] }
        status: { type: string }
        value: { type: string }
        gasPrice: { type: string }
        gasUsed: { type: string }
        gasLimit: { type: string }
        type: { type: string }
        nonce: { type: integer }
        input: { type: string }
        logs:
          type: array
          items: { type: object }
        contractAddress: { type: ["string", "null"] }
        effectiveGasPrice: { type: string }
        cumulativeGasUsed: { type: string }
    Validator:
      type: object
      required: [address, moniker, votingPower, commission, status, uptime, delegators]
      properties:
        address: { type: string }
        moniker: { type: string }
        votingPower: { type: string }
        commission: { type: number }
        status: { type: string }
        uptime: { type: number }
        delegators: { type: integer }
        selfStake: { type: string }
        totalStake: { type: string }
        rank: { type: integer }
        website: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
...
    ValidatorDetail:
      type: object
      required: [address, moniker, votingPower, commission, status, uptime, delegators, selfStake, totalStake, rank]
      properties:
        address: { type: string }
        moniker: { type: string }
        votingPower: { type: string }
        commission: { type: number }
        status: { type: string }
        uptime: { type: number }
        delegators: { type: integer }
        selfStake: { type: string }
        totalStake: { type: string }
        rank: { type: integer }
        website: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        blocksProposed: { type: integer }
        blocksSignedLast100: { type: integer }
        recentBlocks:
          type: array
          items: { type: integer }
...
    Proposal:
      type: object
      required: [id, title, status, proposer, yesVotes, noVotes, abstainVotes, endTime, type]
      properties:
        id: { type: integer }
        title: { type: string }
        status: { type: string }
        proposer: { type: string }
        yesVotes: { type: string }
        noVotes: { type: string }
        abstainVotes: { type: string }
        endTime: { type: string }
        type: { type: string }
        summary: { type: string }
        zepNumber: { type: ["integer", "null"] }
    ProposalDetail:
      type: object
      required: [id, title, status, proposer, yesVotes, noVotes, abstainVotes, endTime, type, description, changes]
      properties:
        id: { type: integer }
        title: { type: string }
        status: { type: string }
        proposer: { type: string }
        yesVotes: { type: string }
        noVotes: { type: string }
        abstainVotes: { type: string }
        endTime: { type: string }
        type: { type: string }
        description: { type: string }
        zepNumber: { type: ["integer", "null"] }
        summary: { type: string }
        changes:
          type: array
          items: { type: string }
        voters:
          type: array
          items:
            type: object
...
    PayIdResolution:
      type: object
      required: [payid, address, network]
      properties:
        payid: { type: string }
        address: { type: string }
        network: { type: string }
        registered: { type: string }
        txCount: { type: integer }
        balance: { type: string }
...