# lib/api-zod/src/generated/types/blockDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ZBX Chain Explorer API
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from './transaction';

export interface BlockDetail {
  number: number;
  hash: string;
  timestamp: string;
  txCount: number;
  proposer: string;
  gasUsed: string;
  gasLimit: string;
  size: number;
  parentHash: string;
  stateRoot: string;
  transactionsRoot: string;
  receiptsRoot: string;
  difficulty: string;
  totalDifficulty: string;
  nonce: string;
  extraData?: string;
  baseFeePerGas?: string;
  transactions?: Transaction[];
}