# lib/api-zod/src/generated/types/transactionDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ZBX Chain Explorer API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionDetailLogsItem } from './transactionDetailLogsItem';

export interface TransactionDetail {
  hash: string;
  blockNumber: number;
  timestamp: string;
  from: string;
  /** @nullable */
  to?: string | null;
  status: string;
  value: string;
  gasPrice: string;
  gasUsed: string;
  gasLimit?: string;
  type: string;
  nonce: number;
  input: string;
  logs: TransactionDetailLogsItem[];
  /** @nullable */
  contractAddress?: string | null;
  effectiveGasPrice?: string;
  cumulativeGasUsed?: string;
}