# lib/api-zod/src/generated/types/validator.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ZBX Chain Explorer API
 * OpenAPI spec version: 0.1.0
 */

export interface Validator {
  address: string;
  moniker: string;
  votingPower: string;
  commission: number;
  status: string;
  uptime: number;
  delegators: number;
  selfStake?: string;
  totalStake?: string;
  rank?: number;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  description?: string | null;
}